tests: rip off test_thread.hpp for finer granularity
Original header is torn to three pieces - `thr_ctx`-related
declarations, standalone functions and threadpool-related declarations,
and inclusion of the library header bringing parallel functions from
there.

The purpose of this division is to get fine granularity of abstractions
or symbols needed in test files to avoid undesired transitive inclusion
of the library headers into tests.
