+++ tests/test_thread.hpp
@@ -14,11 +14,8 @@
 * limitations under the License.
 *******************************************************************************/
 
-#ifndef TEST_THREAD_HPP
-#define TEST_THREAD_HPP
-
-#include <functional>
-#include <iostream>
+#ifndef TESTS_TEST_THREAD_HPP
+#define TESTS_TEST_THREAD_HPP
 
 #include "oneapi/dnnl/dnnl_config.h"
 
@@ -47,22 +44,6 @@
 
 #endif
 
-// Here we define some types in global namespace to handle customized
-// threading context for creation and execution
-struct thr_ctx_t {
-    int max_concurrency;
-    int core_type;
-    int nthr_per_core;
-
-    bool operator==(const thr_ctx_t &rhs) const {
-        return max_concurrency == rhs.max_concurrency
-                && core_type == rhs.core_type
-                && nthr_per_core == rhs.nthr_per_core;
-    }
-    bool operator!=(const thr_ctx_t &rhs) const { return !(*this == rhs); }
-    void *get_interop_obj() const;
-};
-
 // This hack renames the namespaces used by threading functions for
 // threadpool-related functions so that the calls to dnnl::impl::parallel*()
 // from the test use a special testing threadpool.
@@ -84,80 +65,5 @@ struct thr_ctx_t {
 #ifndef COMMON_DNNL_THREAD_HPP
 #error "src/common/dnnl_thread.hpp" has an unexpected header guard
 #endif
-
-std::ostream &operator<<(std::ostream &os, const thr_ctx_t &ctx);
-
-const thr_ctx_t &get_default_thr_ctx();
-
-#if DNNL_CPU_THREADING_RUNTIME == DNNL_RUNTIME_THREADPOOL
-#include "oneapi/dnnl/dnnl_threadpool_iface.hpp"
-namespace dnnl {
-
-// Original threadpool utils are used by the scoped_tp_activation_t and thus
-// need to be re-declared because of the hack above.
-namespace impl {
-namespace threadpool_utils {
-void activate_threadpool(dnnl::threadpool_interop::threadpool_iface *tp);
-void deactivate_threadpool();
-dnnl::threadpool_interop::threadpool_iface *get_active_threadpool();
-int get_max_concurrency();
-} // namespace threadpool_utils
-} // namespace impl
-
-namespace testing {
-
-dnnl::threadpool_interop::threadpool_iface *get_threadpool(
-        const thr_ctx_t &ctx = get_default_thr_ctx());
-
-// Sets the testing threadpool as active for the lifetime of the object.
-// Required for the tests that throw to work.
-struct scoped_tp_activation_t {
-    scoped_tp_activation_t(dnnl::threadpool_interop::threadpool_iface *tp_
-            = get_threadpool()) {
-        impl::threadpool_utils::activate_threadpool(tp_);
-    }
-    ~scoped_tp_activation_t() {
-        impl::threadpool_utils::deactivate_threadpool();
-    }
-};
-
-struct scoped_tp_deactivation_t {
-    scoped_tp_deactivation_t() {
-        impl::threadpool_utils::deactivate_threadpool();
-    }
-    ~scoped_tp_deactivation_t() {
-        // we always use the same threadpool that is returned by `get_threadpool()`
-        impl::threadpool_utils::activate_threadpool(get_threadpool());
-    }
-};
-
-} // namespace testing
-} // namespace dnnl
-#endif
-
-// These are free functions to allow running a function in a given threading
-// context.
-// A threading context is defined by:
-// - number of threads
-// - type of cores (TBB only)
-// - threads per core (TBB only)
-
-// Note: we have to differentiate creation and execution in thread
-// context because of threadpool as it uses different mecanisms in
-// both (in execution, tp is passed in stream)
-//
-// Definitions live in test_thread.cpp where the runtime-specific logic is
-// handled inside a single version of each function.
-
-int create_in_thr_ctx(const thr_ctx_t &ctx, const std::function<int()> &f);
-// The function f shall take an interop obj as last argument
-int execute_in_thr_ctx(const thr_ctx_t &ctx, const std::function<int()> &f);
-
-// TBB runtime may crash when it is used under CTest. This is a known TBB
-// limitation that can be worked around by doing explicit finalization.
-// The API to do that was introduced in 2021.6.0. When using an older TBB
-// runtime the crash may still happen.
-// Appropriate header lives in a `src/common/dnnl_thread_tbb_proxy.hpp`.
-void finalize_tbb();
 
 #endif
